+++ src/snowflake/snowpark/_internal/data_source/drivers/base_driver.py
@@ -3,11 +3,9 @@
 #
 from enum import Enum
 import time
-import datetime
 from typing import List, Callable, Any, Optional, TYPE_CHECKING
 from snowflake.connector.options import pandas as pd
 
-from snowflake.snowpark._internal.analyzer.analyzer_utils import unquote_if_quoted
 from snowflake.snowpark._internal.data_source.datasource_typing import (
     Connection,
     Cursor,
@@ -19,10 +17,6 @@ from snowflake.snowpark.types import (
     StructType,
     StructField,
     VariantType,
-    TimestampType,
-    IntegerType,
-    BinaryType,
-    DateType,
 )
 import snowflake.snowpark
 import logging
@@ -160,35 +154,6 @@ class BaseDriver:
             if get_sorted_key_for_version(str(pd.__version__)) < (2, 1, 0)
             else pandas_df.map
         )
-
-    @staticmethod
-    def data_source_data_to_pandas_df(
-        data: List[Any], schema: StructType
-    ) -> "pd.DataFrame":
-        # unquote column name because double quotes stored in parquet file create column mismatch during copy into table
-        columns = [unquote_if_quoted(col.name) for col in schema.fields]
-        # this way handles both list of object and list of tuples and avoid implicit pandas type conversion
-        df = pd.DataFrame([list(row) for row in data], columns=columns, dtype=object)
-
-        for field in schema.fields:
-            name = unquote_if_quoted(field.name)
-            if isinstance(field.datatype, IntegerType):
-                # 'Int64' is a pandas dtype while 'int64' is a numpy dtype, as stated here:
-                # https://github.com/pandas-dev/pandas/issues/27731
-                # https://pandas.pydata.org/docs/reference/api/pandas.Int64Dtype.html
-                # https://numpy.org/doc/stable/reference/arrays.scalars.html#numpy.int64
-                df[name] = df[name].astype("Int64")
-            elif isinstance(field.datatype, (TimestampType, DateType)):
-                df[name] = df[name].map(
-                    lambda x: x.isoformat()
-                    if isinstance(x, (datetime.datetime, datetime.date))
-                    else x
-                )
-            elif isinstance(field.datatype, BinaryType):
-                df[name] = df[name].map(
-                    lambda x: x.hex() if isinstance(x, (bytearray, bytes)) else x
-                )
-        return df
 
     @staticmethod
     def to_result_snowpark_df(
