+++ src/snowflake/snowpark/_internal/data_source/utils.py
@@ -1,13 +1,15 @@
 #
 # Copyright (c) 2012-2025 Snowflake Computing Inc. All rights reserved.
 #
-
+import datetime
 import os
 import queue
 import traceback
 from enum import Enum
-from typing import Any, Tuple, Optional, Callable, Dict, Set, Union
+from typing import Any, Tuple, Optional, Callable, Dict, Set, Union, List
 import logging
+
+from snowflake.snowpark._internal.analyzer.analyzer_utils import unquote_if_quoted
 from snowflake.snowpark._internal.data_source.dbms_dialects import (
     Sqlite3Dialect,
     OracledbDialect,
@@ -29,7 +31,14 @@ from snowflake.snowpark._internal.data_source import DataSourceReader
 from snowflake.snowpark._internal.type_utils import type_string_to_type_object
 from snowflake.snowpark._internal.utils import normalize_local_file
 from snowflake.snowpark.exceptions import SnowparkDataframeReaderException
-from snowflake.snowpark.types import StructType
+from snowflake.snowpark.types import (
+    StructType,
+    IntegerType,
+    TimestampType,
+    DateType,
+    BinaryType,
+)
+from snowflake.connector.options import pandas as pd
 
 logger = logging.getLogger(__name__)
 
@@ -149,7 +158,7 @@ def _task_fetch_data_from_source(
     tmp_dir: str,
 ):
     def convert_to_parquet(fetched_data, fetch_idx):
-        df = worker.data_source_data_to_pandas_df(fetched_data)
+        df = data_source_data_to_pandas_df(fetched_data, worker.schema)
         if df.empty:
             logger.debug(
                 f"The DataFrame is empty, no parquet file is generated for partition {partition_idx} fetch {fetch_idx}."
@@ -283,3 +292,32 @@ def convert_custom_schema_to_structtype(
             'The schema should be either a valid schema string, for example: "id INTEGER, int_col INTEGER, text_col STRING".'
             'or a valid StructType, for example: StructType([StructField("ID", IntegerType(), False)])'
         )
+
+
+def data_source_data_to_pandas_df(
+    data: List[Any], schema: StructType
+) -> "pd.DataFrame":
+    # unquote column name because double quotes stored in parquet file create column mismatch during copy into table
+    columns = [unquote_if_quoted(col.name) for col in schema.fields]
+    # this way handles both list of object and list of tuples and avoid implicit pandas type conversion
+    df = pd.DataFrame([list(row) for row in data], columns=columns, dtype=object)
+
+    for field in schema.fields:
+        name = unquote_if_quoted(field.name)
+        if isinstance(field.datatype, IntegerType):
+            # 'Int64' is a pandas dtype while 'int64' is a numpy dtype, as stated here:
+            # https://github.com/pandas-dev/pandas/issues/27731
+            # https://pandas.pydata.org/docs/reference/api/pandas.Int64Dtype.html
+            # https://numpy.org/doc/stable/reference/arrays.scalars.html#numpy.int64
+            df[name] = df[name].astype("Int64")
+        elif isinstance(field.datatype, (TimestampType, DateType)):
+            df[name] = df[name].map(
+                lambda x: x.isoformat()
+                if isinstance(x, (datetime.datetime, datetime.date))
+                else x
+            )
+        elif isinstance(field.datatype, BinaryType):
+            df[name] = df[name].map(
+                lambda x: x.hex() if isinstance(x, (bytearray, bytes)) else x
+            )
+    return df
